requirements.txt
langchain-astradb
langchain 
google-generativeai
langchain-google-genai
datasets
pypdf
python-dotenv
flask

-e .